Add database with initial implementation
I chose PostgreSQL and asyncpg.

PostgreSQL because it's simple and well understood even though it might
not be the best choice for storing time series data. But that's why it
is abtracted away and thus can be replaced without a full rewrite.

Asyncpg because it's fast and simple library that doesn't require much
configuration and is optimized for direct DB usage instead of requiring
models and whatnot.